+++ test/validation_hints_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class ValidationHintsTest < Minitest::Test
   def test_version
-    assert_equal "6.3.0", ValidationHints::VERSION
+    assert_equal "6.4.0", ValidationHints::VERSION
   end
 
   def test_locale_path_exists